LayoutTests/svg/animations/svglength-additive-from-by-2-expected.txt

SVG 1.1 dynamic animation tests


This tests from-by-animations adding to previous underlying values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 10
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 0
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 10
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 10
PASS successfullyParsed is true

TEST COMPLETE
